+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Changelog
 
-## TODO
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.2.0] - January, 2024
+
+### Features
+
+* Add an optional temporary directory: `temp_dir` (#28).